// src/entities/openapi/model/signatureSlotCodeEnum.ts
/**
 * agridata.ch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const SignatureSlotCodeEnum = {
    DataConsumer01: 'DATA_CONSUMER_01',
    DataConsumer02: 'DATA_CONSUMER_02',
    DataProvider01: 'DATA_PROVIDER_01',
    DataProvider02: 'DATA_PROVIDER_02'
} as const;
export type SignatureSlotCodeEnum = typeof SignatureSlotCodeEnum[keyof typeof SignatureSlotCodeEnum];